split VirtualRecord off into own header
